+++ Service/PostGalleryManager.php
@@ -1,0 +1,165 @@
+<?php
+
+/**
+ * This file is part of the Bono CMS
+ * 
+ * Copyright (c) No Global State Lab
+ * 
+ * For the full copyright and license information, please view
+ * the license file that was distributed with this source code.
+ */
+
+namespace News\Service;
+
+use Krystal\Stdlib\VirtualEntity;
+use Krystal\Image\Tool\ImageManager;
+use Cms\Service\AbstractManager;
+use News\Storage\PostGalleryMapperInterface;
+
+final class PostGalleryManager extends AbstractManager implements PostGalleryManagerInterface
+{
+    /**
+     * Any compliant post gallery mapper
+     * 
+     * @var \News\Storage\PostGalleryMapperInterface
+     */
+    private $postGalleryMapper;
+
+    /**
+     * Image manager for posts
+     * 
+     * @var \Krystal\Image\Tool\ImageManager
+     */
+    private $imageManager;
+
+    /**
+     * State initialization
+     * 
+     * @param \News\Storage\PostGalleryMapperInterface $postGalleryMapper
+     * @param \Krystal\Image\Tool\ImageManager $imageManager
+     * @return void
+     */
+    public function __construct(PostGalleryMapperInterface $postGalleryMapper, ImageManager $imageManager)
+    {
+        $this->postGalleryMapper = $postGalleryMapper;
+        $this->imageManager = $imageManager;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function toEntity(array $row)
+    {
+        // Configure image bag
+        $imageBag = clone $this->imageManager->getImageBag();
+        $imageBag->setId((int) $row['id'])
+                 ->setCover($row['image']);
+
+        $entity = new ImageEntity();
+        $entity->setId($row['id'])
+               ->setPostId($row['post_id'])
+               ->setOrder($row['order'])
+               ->setImage($row['image'])
+               ->setImageBag($imageBag);
+
+        return $entity;
+    }
+
+    /**
+     * Deletes an image by its associated ID
+     * 
+     * @param int $id Image ID
+     * @return boolean
+     */
+    public function deleteById($id)
+    {
+        return $this->postGalleryMapper->deleteByPk($id) && $this->imageManager->delete($id);
+    }
+
+    /**
+     * Returns last ID
+     * 
+     * @return integer
+     */
+    public function getLastId()
+    {
+        return $this->postGalleryMapper->getMaxId();
+    }
+
+    /**
+     * Fetch images by attached post ID
+     * 
+     * @param int $postId
+     * @return array
+     */
+    public function fetchAllByPostId($postId)
+    {
+        return $this->prepareResults($this->postGalleryMapper->fetchAllByPostId($postId));
+    }
+
+    /**
+     * Finds image by its associated id
+     * 
+     * @param int $id Image ID
+     * @return mixed
+     */
+    public function fetchById($id)
+    {
+        return $this->prepareResult($this->postGalleryMapper->fetchById($id));
+    }
+
+    /**
+     * Adds an image
+     * 
+     * @param array $input
+     * @return boolean
+     */
+    public function add(array $input)
+    {
+        $image = $input['data']['image'];
+        $file = $input['files']['file'];
+
+        // Filter files input
+        $this->filterFileInput($file);
+
+        // Define image attribute
+        $image['image'] = $file[0]->getName();
+
+        // Save image first, because we need to get its ID for image uploading
+        $this->postGalleryMapper->persist($image);
+
+        // And now upload image
+        $this->imageManager->upload($this->getLastId(), $file);
+
+        return true;
+    }
+
+    /**
+     * Updates an image
+     * 
+     * @param array $input
+     * @return boolean
+     */
+    public function update(array $input)
+    {
+        // Grab a reference to image data
+        $image = $input['data']['image'];
+
+        // If file new provided, than start handling
+        if (!empty($input['files'])) {
+            // If we have a previous cover, then we gotta remove it
+            $this->imageManager->delete($image['id'], $image['image']);
+
+            $file = $input['files']['file'];
+
+            // Before we start uploading a file, we need to filter its base name
+            $this->filterFileInput($file);
+            $this->imageManager->upload($image['id'], $file);
+
+            // Now override cover's value with file's base name we currently have from user's input
+            $image['image'] = $file[0]->getName();
+        }
+
+        return $this->postGalleryMapper->persist($image);
+    }
+}